+++ emailBackEnd/emails/munpack/codes.c
@@ -136,7 +136,7 @@ long *len;
     int nbytes;
     
     MD5Init(&context);
-    while (nbytes = fread(buf, 1, sizeof(buf), infile)) {
+    while (nbytes = (int)fread(buf, 1, sizeof(buf), infile)) {
 	length += nbytes;
 	MD5Update(&context, buf, nbytes);
     }
